日增量变更数据入库——笔记点赞/取消点赞

+++ xiaochengshu-data-align/src/main/java/com/chengliuxiang/xiaochengshu/data/align/constant/TableConstants.java
@@ -13,7 +13,7 @@ public class TableConstants {
      * @param hashKey
      * @return
      */
-    public static String buildTableNameSuffix(String date, int hashKey) {
+    public static String buildTableNameSuffix(String date, long hashKey) {
         return date + TABLE_NAME_SEPARATE + hashKey;
     }
 }
